requirements.txt
# Core dependencies
python-telegram-bot==20.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
pytz==2023.3
httpx==0.25.1
aiofiles==23.2.1
qrcode[pil]==7.4.2
Pillow==10.1.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0